Use setuptools.find_packages
Defining the packages manually would work if you run the build
sequence manually, however this becomes a problem when building with
`pip wheel .`. `pip`'s first operation is to run the `egg_info`
subcommand to discover basic information (i.e. name & version). At
this point the package information is incorrect, but it doesn't
matter because it'll build dist-info in a later step.