Add W503 to ignored flake8 warnings
The reason for adding W503 to the ignored list is because W503 (Line
break occurred before a binary operator) and W504 (Line break occurred
after a binary operator) are both enabled in flake8 by default for some
reason.  But they are mutually exclusive so one has to be disabled /
ignored.